# docs/swagger.yaml
definitions:
  model.Transaction:
    properties:
      amount:
        type: integer
      category:
        type: string
      date:
        type: string
      group:
        type: string
      id:
        type: string
      paidBy:
        type: string
      paidFor:
        items:
          $ref: '#/definitions/model.TransactionTarget'
        type: array
      title:
        type: string
    type: object
  model.TransactionTarget:
    properties:
      forcePrice:
        type: integer
      user:
        type: string
      weight:
        type: integer
    type: object
info:
  contact: {}
  description: Triplan API POC
  title: Triplan API
  version: "1.0"
paths:
  /groups/{id}/transactions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: The transaction to create
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.Transaction'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
      summary: Creates a transaction
  /transactions/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
      summary: Deletes a transaction
    put:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: The transaction to update
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.Transaction'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
      summary: Updates a transaction
  /trips/{id}/transactions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
      summary: Returns all the spending from this trip
swagger: "2.0"